5/dbms/lab/lab-11/requirements.txt
psycopg2-binary
sentence-transformers
python-dotenv
pgvector
